Set detection threshold in constructor

+++ pj_tflite_face_dbface/image_processor/face_detection_engine.cpp
@@ -279,7 +279,7 @@ int32_t FaceDetectionEngine::Process(const cv::Mat& original_mat, Result& result
 float FaceDetectionEngine::ExpSpecial(float x)
 {
     static constexpr int32_t gate = 1;
-    static const float base = std::exp(1);
+    static const float base = static_cast<float>(std::exp(1));
     if (std::abs(x) < gate) {
         return x * base;
     } else if (x > 0) {
